+ [EMBEDU-317] Enable kernel modules debug at module entry/exit
Kernel source:
'scripts/gdb/linux/symbols.py' does not load debug symbols for sections
".init.text" and ".exit.text". So gdb does not know about debug symbols
of this sections. Let's load these symbols, so we could know __init function's
names and line numbers in gdb.

Also,
this patch enables gdb to break at module's __init.
It's not possible to do with "b init_function" command, as module is not
loaded yet, so we don't know the address to put a breakpoint.

To break at __init we first break to 'do_init_module', which alreafy knows
the location of module's __init and __exit functions, and so we can also
set breakpoints from it with 'b mod->init' / 'b mod->exit'